requirements/base.txt
Flask==2.1.2
bootstrap-flask==2.0.2
flask-sqlalchemy==2.5.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
flask-wtf==1.0.1